pyproject.toml
[project]
name = "agent-research"
version = "1.0.0"
description = "Optimized deep research agent using Agno 1.7.0 with structured outputs and citation support"
readme = "README.md"
requires-python = ">=3.10"
authors = [{name = "Agent Research Team"}]
license = {text = "MIT"}
keywords = ["ai", "research", "agents", "openai", "agno", "citations"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "agno>=1.7.0",
    "openai>=1.92.2",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "rich>=14.0.0",
]

[project.urls]
Homepage = "https://github.com/user/agent-research"
Repository = "https://github.com/user/agent-research.git"
Documentation = "https://github.com/user/agent-research#readme"
Issues = "https://github.com/user/agent-research/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]